Isi utama sejak merge terakhir:

· Printer (laporan lapangan, dua putaran): IP yang diketik tak lagi hilang saat
  layar ditutup, kegagalan menyimpan tak lagi diam, dan alamat kosong/salah
  ketik tak lagi berjalan lurus ke soket — "Failed host lookup: ''" berubah
  jadi "IP printer belum diisi", seketika.
· Toleransi yang DIUKUR: `1e-9` berhenti berarti di besaran 10^7, dan menolak
  penjualan yang stoknya persis cukup (400 "sisa 245.785, butuh 245.785").
· Saldo rak yang disusun di JS: sisa yang ADA tak bisa dihabiskan
  (0.19999999999999998) — dikembalikan ke presisi kolomnya.
· Replay antrean offline memotong stok DUA KALI (100→93→86); enam situs kini
  membagi kunci idempotensi dengan percobaan online-nya.
· Ledger sinkron, tiga tabel token, dan berkas unggahan yatim: yang tumbuh
  selamanya kini punya retensi.
· Sel CSV yang dieksekusi Excel dinetralkan; badan permintaan menolak kunci
  asing; angka yang meluap dibalas 400 bernama.

CHANGELOG-API: 13 entri distempel "sudah di-merge ke production".

Gerbang dijalankan pada HASIL MERGE ini sebelum didorong, sesuai prosedur di
kepala ci.yml: typecheck bersih · npm test 2.261 (192 berkas) · build web ·
verify-api 3.006 lolos 0 gagal vs Postgres SEGAR · cakupan rute 272 identik
dengan rekaman · audit invarian 26/26 · Playwright e2e 6/6.

Prosedur itu terbayar: e2e ternyata MERAH pada hasil merge (§248 mengganti
nama cabang seed dan tak mengembalikannya, sementara e2e membaca prefiks
nomor struk). Diperbaiki lebih dulu — tanpa itu job deploy akan tertahan
tanpa satu pun uji unit berubah warna.

+++ apps/server/src/lib/porsi-ditagih.ts
@@ -15,12 +15,25 @@
  * `GREATEST(…, 0)` bukan sekadar jaga-jaga: `qtyDitagih` di shared juga
  * menjepit ke nol, dan dua sisi yang sama-sama menjepit tak bisa berbeda tanda
  * kalau suatu hari ada data yang `qty_refund`-nya melampaui `qty`.
+ *
+ * SEMUANYA BERAKHIR `::float8`, DAN ITU BUKAN KOSMETIK. `sql<number>` adalah
+ * janji yang ditulis manusia dan tak diperiksa siapa pun: driver `pg`
+ * memulangkan `numeric` — termasuk hasil `SUM()`-nya — sebagai STRING.
+ * Tanpa cast, `sql<number>` bertipe `number` di TypeScript dan berisi
+ * `"3000.75"` saat dijalankan; `a + b` merangkainya jadi `"3000.753000.75"`,
+ * dan nilainya yang lolos ke `c.json()` melanggar kontrak API sebagai string
+ * di tempat klien menunggu angka (`as num` di Dart melempar).
+ *
+ * Cast ini TIDAK mengubah satu angka pun: `x::float8` menghasilkan float64
+ * yang bit-per-bit sama dengan `Number("x")`, yang memang sudah dipakai tiap
+ * pemanggil. Yang berubah cuma di mana kebenarannya tinggal — di ekspresinya
+ * sendiri, bukan pada ingatan tiap orang yang kelak memakainya.
  */
 import { sql } from "drizzle-orm";
 import { saleItems } from "../db/schema";
 
 /** `qty − qty_refund`, minimal 0. */
-export const qtyDitagihSql = sql<number>`GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)`;
+export const qtyDitagihSql = sql<number>`GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)::float8`;
 
 /**
  * Omzet kotor baris SESUDAH refund.
@@ -31,10 +44,10 @@ export const qtyDitagihSql = sql<number>`GREATEST(${saleItems.qty} - ${saleItems
  * `sales.subtotal` sesudah refund — `hitungUangSetelahRefund` menyusunnya dari
  * penjumlahan yang sama.
  */
-export const omzetDitagihSql = sql<number>`COALESCE(SUM(${saleItems.hargaSatuan} * GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)), 0)`;
+export const omzetDitagihSql = sql<number>`COALESCE(SUM(${saleItems.hargaSatuan} * GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)), 0)::float8`;
 
 /** Jumlah porsi yang ditagih pada satu grup agregasi. */
-export const sumQtyDitagihSql = sql<number>`COALESCE(SUM(GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)), 0)`;
+export const sumQtyDitagihSql = sql<number>`COALESCE(SUM(GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)), 0)::float8`;
 
 /** HPP yang ditagih: `hpp_satuan` tetap per porsi, dikalikan porsi yang jadi. */
-export const hppDitagihSql = sql<number>`COALESCE(SUM(${saleItems.hppSatuan} * GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)), 0)`;
+export const hppDitagihSql = sql<number>`COALESCE(SUM(${saleItems.hppSatuan} * GREATEST(${saleItems.qty} - ${saleItems.qtyRefund}, 0)), 0)::float8`;
